apps/ml/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["Manik Rana <manikrana54@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.1.0"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
torch = "^2.6.0"
torchvision = "^0.21.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu/torch_stable.html"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
